+++ src/main/java/sehee/answer/Answer.java
@@ -1,4 +1,4 @@
-package sehee.domain;
+package sehee.answer;
 
 import static sehee.util.constant.RandomNumberGameConstant.ANSWER_LENGTH;
